# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amane API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/feeds": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List Feeds",
        "operationId": "list_feeds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "feeds"
        ],
        "summary": "Create Feed",
        "operationId": "create_feed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/items": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List All Feed Items",
        "description": "跨源条目列表. feed_id 优先于 group; group 空串只含未分组源, 非空为前缀匹配.",
        "operationId": "list_all_feed_items",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FeedItemState",
              "default": "active"
            }
          },
          {
            "name": "feed_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feed Id"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/{feed_id}": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "Get Feed",
        "operationId": "get_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "feeds"
        ],
        "summary": "Update Feed",
        "operationId": "update_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "feeds"
        ],
        "summary": "Delete Feed",
        "operationId": "delete_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/{feed_id}/poll": {
      "post": {
        "tags": [
          "feeds"
        ],
        "summary": "Poll Feed",
        "operationId": "poll_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/{feed_id}/items": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List Feed Items",
        "operationId": "list_feed_items",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FeedItemState",
              "default": "active"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/feeds/{feed_id}/items/batch": {
      "post": {
        "tags": [
          "feeds"
        ],
        "summary": "Batch Feed Items",
        "operationId": "batch_feed_items",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feed Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedItemBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedItemBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CacheKind": {
        "type": "string",
        "enum": [
          "metadata",
          "trans"
        ],
        "title": "CacheKind",
        "description": "刮削可复用的缓存种类. use_cache 为其集合: 含某项 = 该缓存生效, 不含 = 强制刷新该项."
      },
...
      "ContentType": {
        "type": "string",
        "enum": [
          "censored",
          "uncensored",
          "chinese",
          "western",
          "fc2",
          "amateur",
          "hentai"
        ],
        "title": "ContentType"
      },
...
      "FeedCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "group": {
            "type": "string",
            "title": "Group",
            "default": ""
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "auto_enqueue": {
            "type": "boolean",
            "title": "Auto Enqueue",
            "default": true
          },
          "interval_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Interval Seconds",
            "default": 3600
          },
          "number_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Pattern"
          },
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Use Cache"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "FeedCreateRequest"
      },
...
      "FeedItemBatchRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/FeedItemBatchAction"
          },
          "ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "action",
          "ids"
        ],
        "title": "FeedItemBatchRequest"
      },
      "FeedItemBatchResponse": {
        "properties": {
          "affected": {
            "type": "integer",
            "title": "Affected"
          },
          "missing": {
            "type": "integer",
            "title": "Missing"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          },
          "submitted": {
            "type": "integer",
            "title": "Submitted",
            "default": 0
          },
          "task_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Task Ids"
          }
        },
        "type": "object",
        "required": [
          "affected",
          "missing"
        ],
        "title": "FeedItemBatchResponse"
      },
      "FeedItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "FeedItemListResponse"
      },
      "FeedItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "feed_id": {
            "type": "integer",
            "title": "Feed Id"
          },
          "item_key": {
            "type": "string",
            "title": "Item Key"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "ignored_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ignored At"
          },
          "metadata_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "feed_id",
          "item_key",
          "created_at"
        ],
        "title": "FeedItemResponse"
      },
      "FeedItemState": {
        "type": "string",
        "enum": [
          "active",
          "ignored",
          "all"
        ],
        "title": "FeedItemState",
        "description": "Feed 历史列表状态过滤."
      },
      "FeedListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "FeedListResponse"
      },
      "FeedResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "group": {
            "type": "string",
            "title": "Group",
            "default": ""
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "auto_enqueue": {
            "type": "boolean",
            "title": "Auto Enqueue"
          },
          "interval_seconds": {
            "type": "integer",
            "title": "Interval Seconds"
          },
          "number_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Pattern"
          },
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "title": "Use Cache"
          },
          "next_fetch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Fetch At"
          },
          "last_fetched_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fetched At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_enqueued": {
            "type": "integer",
            "title": "Last Enqueued",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "url",
          "enabled",
          "auto_enqueue",
          "interval_seconds"
        ],
        "title": "FeedResponse"
      },
      "FeedUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "auto_enqueue": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Enqueue"
          },
          "interval_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Seconds"
          },
          "number_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Pattern"
          },
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "use_cache": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Cache"
          }
        },
        "type": "object",
        "title": "FeedUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}